rename: ChatPage -> ChatBox, ChatBox -> ChatBoxContent
