+++ src/components/ChatBox/index.ts
@@ -1,2 +1,1 @@
-export * from './ChatBox';
 export { default } from './ChatBox';
