+++ src/components/ChatBoxContent/index.ts
@@ -1,0 +1,2 @@
+export * from './ChatBoxContent';
+export { default } from './ChatBoxContent';
